pyproject.toml
[tool.poetry]
name = "mir"
version = "1.0.0"
description = "MIR ML Engineering Challenge - CSV to PostgreSQL loader"
authors = ["Developer <dev@mir.com>"]
readme = "README.md"
packages = [
    {include = "app"},
    {include = "dataset"}
]

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
pandas = "^2.3.2"
numpy = { version = "^2.3.2", python = ">=3.11" }
psycopg2-binary = "^2.9.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^1.4.49"
requests = "^2.32.5"
loguru = "^0.7.3"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
openai = "^1.30.0"
httpx = "^0.27.0"
pydantic = "^2.0.0"
jinja2 = "^3.1.0"
apache-airflow = "^3.0.4"
fastapi = "^0.116.1"
mlflow = "^3.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
init-db = "dataset.init_database:main"
load-csv = "dataset.load_csv_to_postgres:main"
analyze-data = "dataset.analyze_data:main"
drop-db = "dataset.drop_all_tables:main"
serve-api = "app.server:main"
setup-db = "dataset.setup_database:main"
